pyproject.toml
[tool.poetry]
name = "performance-engineering-with-gpt"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
PyGithub = "^1.58.1"
python = "^3.10poe"
ipykernel = "^6.22.0"
pandas = "^2.0.1"
openpyxl = "^3.1.2"
[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
